Bump to latest Elastic.Xunit and Proc (#3383)
Clusters can now quite easily provide an expected shutdown period.

Proc waits for the streamreaders to emit EOL after the process has
exitted and if this does not happen before a timeout OnError on the
observable is called.

In Elastic.Xunit we call `elasticsearch.bat` which spawns
`elasticsearch.jar` and so `ElasticsearchNode` tracks both process id's.

So while Proc was shutting down the host OK as that triggers the read
wait for EOL on the console out streams these pipes (coming from
`elasticsearch.jar`) can stay active quite a while. When we give up the
jar is kept running (in shutdown mode).

This manifested itself later on when `EphemeralClusterComposer` wanted
to run its clean up routines and the ephemeral directories were still in
use.

NodeConfiguration can now provide `WaitForShutDown` which will set
`Proc`'s `WaitForExit` and `WaitForStreamReaders`. `ClusterBase` also
provided an overridable `ModifyNodeConfiguration` so that this can be
set per cluster with relatively low ceremony.

This `WaitForShutdown` does not have to be accurate if its too short
`ElasticsearchNode` now does a hard kill on both the host and and the
java pid in a new hook provided by `ObservableProcessBase`
`OnBeforeWaitForEndOfStreamsError`.
Conflicts:
	src/Tests/Tests.Core/Tests.Core.csproj
	src/Tests/Tests.Domain/Tests.Domain.csproj

Cherry-picked from b0536b9